refactor(components): 👷 folders structure
