+++ app/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next'
 
+import PageNavigation from '@/components/layout/page-navigation/PageNavigation'
 import PageContainer from '@/components/layout/PageContainer'
-import PageNavigation from '@/components/page-navigation/PageNavigation'
 import Companies from '@/components/pages/home/Companies'
 import WhatIDoMindset from '@/components/pages/home/expertise/MySKills'
 import Hero from '@/components/pages/home/hero/Hero'
@@ -9,7 +9,7 @@ import ScanMyContactQR from '@/components/pages/home/ScanMyContactQR'
 import Skills from '@/components/pages/home/skills/Skills'
 import SkillsMain from '@/components/pages/home/skills/SkillsMain'
 
-import { metaDataShared } from '@/lib/data/metadata/metaDataShared'
+import { metaDataShared } from '@/lib/data/metadata/shared/metaDataShared'
 
 import { TEXT } from '@/localization/english'
