+++ app/testimonials/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next'
 
 import PageContainer from '@/components/layout/PageContainer'
-import PageNavigation from '@/components/page-navigation/PageNavigation'
+import PageNavigation from '@/components/layout/page-navigation/PageNavigation'
 import BreadCrumbs from '@/components/shared/Breadcrumbs'
 import Heading1 from '@/components/shared/Heading1'
 import Highlight from '@/components/shared/Highlight'
