+++ components/layout/header/ScrollProgressBar.tsx
