+++ components/pages/about-me/Photo.tsx
