+++ lib/data/metadata/shared/appleTouchIconSizes.ts
